statistics: rename queue files for consistency

fix

fix

fix
