+++ pkg/statistics/handle/autoanalyze/priorityqueue/BUILD.bazel
@@ -10,8 +10,7 @@ go_library(
         "job.go",
         "non_partitioned_table_analysis_job.go",
         "queue.go",
-        "queue_v2.go",
-        "queue_v2_ddl_handler.go",
+        "queue_ddl_handler.go",
         "static_partitioned_table_analysis_job.go",
     ],
     importpath = "github.com/pingcap/tidb/pkg/statistics/handle/autoanalyze/priorityqueue",
@@ -50,8 +49,8 @@ go_test(
         "job_test.go",
         "main_test.go",
         "non_partitioned_table_analysis_job_test.go",
-        "queue_v2_ddl_handler_test.go",
-        "queue_v2_test.go",
+        "queue_ddl_handler_test.go",
+        "queue_test.go",
         "static_partitioned_table_analysis_job_test.go",
     ],
     flaky = True,
