+++ pkg/statistics/handle/autoanalyze/priorityqueue/queue_ddl_handler.go
@@ -34,7 +34,7 @@ import (
 )
 
 // HandleDDLEvent handles DDL events for the priority queue.
-func (pq *AnalysisPriorityQueueV2) HandleDDLEvent(_ context.Context, sctx sessionctx.Context, event *notifier.SchemaChangeEvent) (err error) {
+func (pq *AnalysisPriorityQueue) HandleDDLEvent(_ context.Context, sctx sessionctx.Context, event *notifier.SchemaChangeEvent) (err error) {
 	pq.syncFields.mu.Lock()
 	defer pq.syncFields.mu.Unlock()
 	// If the priority queue is not initialized, we should retry later.
@@ -79,7 +79,7 @@ func (pq *AnalysisPriorityQueueV2) HandleDDLEvent(_ context.Context, sctx sessio
 }
 
 // getAndDeleteJob tries to get a job from the priority queue and delete it if it exists.
-func (pq *AnalysisPriorityQueueV2) getAndDeleteJob(tableID int64) error {
+func (pq *AnalysisPriorityQueue) getAndDeleteJob(tableID int64) error {
 	job, ok, err := pq.syncFields.inner.GetByKey(tableID)
 	if err != nil {
 		statslogutil.StatsLogger().Error(
@@ -105,7 +105,7 @@ func (pq *AnalysisPriorityQueueV2) getAndDeleteJob(tableID int64) error {
 }
 
 // recreateAndPushJob is a helper function that recreates a job and pushes it to the queue.
-func (pq *AnalysisPriorityQueueV2) recreateAndPushJob(
+func (pq *AnalysisPriorityQueue) recreateAndPushJob(
 	sctx sessionctx.Context,
 	lockedTables map[int64]struct{},
 	pruneMode variable.PartitionPruneMode,
@@ -113,7 +113,7 @@ func (pq *AnalysisPriorityQueueV2) recreateAndPushJob(
 ) error {
 	parameters := exec.GetAutoAnalyzeParameters(sctx)
 	autoAnalyzeRatio := exec.ParseAutoAnalyzeRatio(parameters[variable.TiDBAutoAnalyzeRatio])
-	currentTs, err := getStartTs(sctx)
+	currentTs, err := statsutil.GetStartTS(sctx)
 	if err != nil {
 		return errors.Trace(err)
 	}
@@ -127,7 +127,7 @@ func (pq *AnalysisPriorityQueueV2) recreateAndPushJob(
 // For static partitioned tables, we need to recreate the job for each partition.
 // So we need to call this function for each partition.
 // For normal tables and dynamic partitioned tables, we only need to recreate the job for the whole table.
-func (pq *AnalysisPriorityQueueV2) recreateAndPushJobForTable(sctx sessionctx.Context, tableInfo *model.TableInfo) error {
+func (pq *AnalysisPriorityQueue) recreateAndPushJobForTable(sctx sessionctx.Context, tableInfo *model.TableInfo) error {
 	pruneMode := variable.PartitionPruneMode(sctx.GetSessionVars().PartitionPruneMode.Load())
 	partitionInfo := tableInfo.GetPartitionInfo()
 	lockedTables, err := lockstats.QueryLockedTables(statsutil.StatsCtx, sctx)
@@ -149,7 +149,7 @@ func (pq *AnalysisPriorityQueueV2) recreateAndPushJobForTable(sctx sessionctx.Co
 	return pq.recreateAndPushJob(sctx, lockedTables, pruneMode, stats)
 }
 
-func (pq *AnalysisPriorityQueueV2) handleAddIndexEvent(
+func (pq *AnalysisPriorityQueue) handleAddIndexEvent(
 	sctx sessionctx.Context,
 	event *notifier.SchemaChangeEvent,
 ) error {
@@ -168,7 +168,7 @@ func (pq *AnalysisPriorityQueueV2) handleAddIndexEvent(
 	parameters := exec.GetAutoAnalyzeParameters(sctx)
 	autoAnalyzeRatio := exec.ParseAutoAnalyzeRatio(parameters[variable.TiDBAutoAnalyzeRatio])
 	// Get current timestamp from the session context.
-	currentTs, err := getStartTs(sctx)
+	currentTs, err := statsutil.GetStartTS(sctx)
 	if err != nil {
 		return errors.Trace(err)
 	}
@@ -198,7 +198,7 @@ func (pq *AnalysisPriorityQueueV2) handleAddIndexEvent(
 	return pq.pushWithoutLock(job)
 }
 
-func (pq *AnalysisPriorityQueueV2) handleTruncateTableEvent(
+func (pq *AnalysisPriorityQueue) handleTruncateTableEvent(
 	_ sessionctx.Context,
 	event *notifier.SchemaChangeEvent,
 ) error {
@@ -224,7 +224,7 @@ func (pq *AnalysisPriorityQueueV2) handleTruncateTableEvent(
 	return nil
 }
 
-func (pq *AnalysisPriorityQueueV2) handleDropTableEvent(
+func (pq *AnalysisPriorityQueue) handleDropTableEvent(
 	_ sessionctx.Context,
 	event *notifier.SchemaChangeEvent,
 ) error {
@@ -250,7 +250,7 @@ func (pq *AnalysisPriorityQueueV2) handleDropTableEvent(
 	return nil
 }
 
-func (pq *AnalysisPriorityQueueV2) handleTruncateTablePartitionEvent(
+func (pq *AnalysisPriorityQueue) handleTruncateTablePartitionEvent(
 	sctx sessionctx.Context,
 	event *notifier.SchemaChangeEvent,
 ) error {
@@ -273,7 +273,7 @@ func (pq *AnalysisPriorityQueueV2) handleTruncateTablePartitionEvent(
 	return pq.recreateAndPushJobForTable(sctx, globalTableInfo)
 }
 
-func (pq *AnalysisPriorityQueueV2) handleDropTablePartitionEvent(
+func (pq *AnalysisPriorityQueue) handleDropTablePartitionEvent(
 	sctx sessionctx.Context,
 	event *notifier.SchemaChangeEvent,
 ) error {
@@ -296,7 +296,7 @@ func (pq *AnalysisPriorityQueueV2) handleDropTablePartitionEvent(
 	return pq.recreateAndPushJobForTable(sctx, globalTableInfo)
 }
 
-func (pq *AnalysisPriorityQueueV2) handleExchangeTablePartitionEvent(
+func (pq *AnalysisPriorityQueue) handleExchangeTablePartitionEvent(
 	sctx sessionctx.Context,
 	event *notifier.SchemaChangeEvent,
 ) error {
@@ -337,7 +337,7 @@ func (pq *AnalysisPriorityQueueV2) handleExchangeTablePartitionEvent(
 	return nil
 }
 
-func (pq *AnalysisPriorityQueueV2) handleReorganizePartitionEvent(
+func (pq *AnalysisPriorityQueue) handleReorganizePartitionEvent(
 	sctx sessionctx.Context,
 	event *notifier.SchemaChangeEvent,
 ) error {
@@ -363,7 +363,7 @@ func (pq *AnalysisPriorityQueueV2) handleReorganizePartitionEvent(
 	return pq.recreateAndPushJobForTable(sctx, globalTableInfo)
 }
 
-func (pq *AnalysisPriorityQueueV2) handleAlterTablePartitioningEvent(sctx sessionctx.Context, event *notifier.SchemaChangeEvent) error {
+func (pq *AnalysisPriorityQueue) handleAlterTablePartitioningEvent(sctx sessionctx.Context, event *notifier.SchemaChangeEvent) error {
 	oldSingleTableID, newGlobalTableInfo, _ := event.GetAddPartitioningInfo()
 
 	// For non-partitioned tables.
@@ -384,7 +384,7 @@ func (pq *AnalysisPriorityQueueV2) handleAlterTablePartitioningEvent(sctx sessio
 	return pq.recreateAndPushJobForTable(sctx, newGlobalTableInfo)
 }
 
-func (pq *AnalysisPriorityQueueV2) handleRemovePartitioningEvent(sctx sessionctx.Context, event *notifier.SchemaChangeEvent) error {
+func (pq *AnalysisPriorityQueue) handleRemovePartitioningEvent(sctx sessionctx.Context, event *notifier.SchemaChangeEvent) error {
 	oldTblID, newSingleTableInfo, droppedPartInfo := event.GetRemovePartitioningInfo()
 
 	// For static partitioned tables.
